libs/evalgen/pyproject.toml
[build-system]
requires = ["hatchling", "hatch-una>=0.5.1"]
build-backend = "hatchling.build"

[project]
name = "evalgen"
version = "0.0.1"
requires-python = "~=3.11"
description = "Evalgen library for the Data Platform"
dependencies = [
    "databricks-sdk>=0.48.0",
    "databricks-sqlalchemy>=2.0.6",
    "fastcore>=1.8.0",
    "fastsql>=2.0.3",
    "monsterui>=1.0.21",
    "polars>=1.12.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "python-fasthtml>=0.12.19",
    "databricks-connect>=16.1.3",
    "databricks-sdk>=0.48.0",
    "databricks-connect>=15.4.8",
    "mlflow>=2.22.0",
    "langchain>=0.3.25",
    "databricks-langchain>=0.4.0",
    "scikit-learn>=1.6.1",
    "databricks-agents>=0.17.2",
    "langchain-community>=0.3.21",
    "sqlmodel>=0.0.24",
]

[project.optional-dependencies]
databricks = ["alembic>=1.16.1"]

[dependency-groups]
dev = [
    "basedpyright>=1.29.0",
    "databricks-sdk>=0.48.0",
    "databricks-connect~=16.1.3",
    "pydantic~=2.11.3",
    "ruff>=0.11.6",
]


[tool.hatch.build.hooks.una-build]
[tool.hatch.metadata.hooks.una-meta]
[tool.uv]
default-groups = ["dev"]
[tool.uv.sources]

[tool.basedpyright]
reportInvalidTypeForm = "none"
reportAssignmentType = "none"
reportGeneralTypeIssues = "none"
reportOptionalMemberAccess = "none"
reportWildcardImportFromLibrary = "none"
pythonVersion = "3.11"
reportUnusedImport = "none"
reportUnknownMemberType = false
reportUnknownVariableType = false
reportAttributeAccessIssue = "hint"
reportUnknownArgumentType = "none"
reportImplicitOverride = "none"
reportRedeclaration = "none"
reportUnusedCallResult = "none"
reportExplicitAny = "none"
reportUnannotatedClassAttribute = "none"
reportUninitializedInstanceVariable = "warning"
reportAny = "none"
